changing behaviour of external links
If a link to an external page is in an app, it will open in safari by
default. This is a change to current behaviour (a link will hijack the
app unless it is set to target="somethingnew").

The user can override this by specifying hosts in their PhoneGap.plist
file that can serve pages the app will load -- this makes more sense to
my mind than opening anything, or never opening anything.