refs #141 VoiceOver not speaking audio view
No longer set the AudioSession to recording type until press the record button.  That mode was preventing VoiceOver from speaking when view first loaded.  Additional accessibility cleanup:  When the audio view loads was posting an UIAccessibilityLayoutChangedNotification.  It really should be a UIAccessibilityScreenChangedNotification since we are loading a new view.  No need to call setIsAccessibilityElement:YES on standard UI controls.