fixes issue #212 media play not always calling success callback
updated to use mediaId to store/retrieve audioFile in soundCache dictionary rather than url (url is not available with all media types). Because I was not always finding the audioFile in audioDidFinishPlaying the success callback was not getting called even though the media did successfully play.
Was returning success from stopAudioRecord but should not be - this resulted in two success callbacks from record.  Now only call success from audioDidFinishRecording.