fix #265 display contact not restoring after pause
The ABPersonViewController has no UI to dismiss.  Adding navigationItems to it does not work properly if the app is sent to the background.  I created a blank ViewController in front of the ABPersonViewController so now the display contact page has a Back button rather than a done button. When back is pressed the entire NavigationController hierarchy is removed.  The back button remains visible on the contact page after the app is restored from going into the background with this view visible.